Move slice-over-set validator test to tests/py312
``slice`` objects only became hashable in Python 3.12, so ``{slice(...)}``
cannot be constructed on the still-supported 3.11 (and 3.10 on main): the test
raised ``TypeError`` at set construction before the validator ran, and pyright
flagged the set literals as unhashable. Relocate the set-specific case to
``tests/py312`` (excluded from pre-3.12 collection and from pyright); the dict
mapping case, which needs no hashable slices, stays with the other slice tests.

Addresses Copilot review feedback on #891.
